test: JwtTokenProvider 테스트 케이스 추가

+++ src/main/java/com/openmpy/taleswiki/auth/jwt/JwtTokenProvider.java
@@ -62,6 +62,6 @@ public class JwtTokenProvider {
         }
 
         final Map<String, Object> payload = getPayload(token);
-        return (Long) payload.get(ID_KEY);
+        return ((Integer) payload.get(ID_KEY)).longValue();
     }
 }
